Move TaskType to utils/types
