+++ src/anomalib/__init__.py
@@ -3,4 +3,9 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+
+from anomalib.utils.types import TaskType
+
+__all__ = ["TaskType"]
+
 __version__ = "1.0.0dev"
