+++ src/anomalib/data/base/depth.py
@@ -8,7 +8,7 @@ import cv2
 import numpy as np
 import torch
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base.dataset import AnomalibDataset
 from anomalib.data.utils import masks_to_boxes, read_depth_image, read_image
