+++ src/anomalib/data/base/video.py
@@ -9,7 +9,7 @@ import albumentations as A  # noqa: N812
 import torch
 from pandas import DataFrame
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base.datamodule import AnomalibDataModule
 from anomalib.data.base.dataset import AnomalibDataset
 from anomalib.data.utils import ValSplitMode, masks_to_boxes
