+++ src/anomalib/data/depth/folder_3d.py
@@ -12,7 +12,7 @@ from pathlib import Path
 import albumentations as A  # noqa: N812
 from pandas import DataFrame, isna
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base import AnomalibDataModule, AnomalibDepthDataset
 from anomalib.data.utils import (
     DirType,
