+++ src/anomalib/data/depth/mvtec_3d.py
@@ -28,7 +28,7 @@ from pathlib import Path
 import albumentations as A  # noqa: N812
 from pandas import DataFrame
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base import AnomalibDataModule, AnomalibDepthDataset
 from anomalib.data.utils import (
     DownloadInfo,
