+++ src/anomalib/data/utils/synthetic.py
@@ -20,7 +20,7 @@ import pandas as pd
 from albumentations.pytorch import ToTensorV2
 from pandas import DataFrame, Series
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base.dataset import AnomalibDataset
 from anomalib.data.utils import Augmenter, Split, read_image
