+++ src/anomalib/data/video/avenue.py
@@ -26,7 +26,7 @@ import numpy as np
 import scipy.io
 from pandas import DataFrame
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
 from anomalib.data.utils import (
