+++ src/anomalib/data/video/ucsd_ped.py
@@ -15,7 +15,7 @@ import numpy as np
 import torch
 from pandas import DataFrame
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.data.base import AnomalibVideoDataModule, AnomalibVideoDataset
 from anomalib.data.base.video import VideoTargetFrame
 from anomalib.data.utils import (
