+++ src/anomalib/utils/callbacks/visualizer/utils.py
@@ -3,7 +3,7 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.post_processing import VisualizationMode
 
 from .visualizer_base import BaseVisualizerCallback
