+++ src/anomalib/utils/callbacks/visualizer/visualizer_base.py
@@ -11,7 +11,7 @@ import lightning.pytorch as pl
 import numpy as np
 from lightning.pytorch import Callback
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.models.components import AnomalyModule
 from anomalib.post_processing import VisualizationMode, Visualizer
 from anomalib.utils.loggers import AnomalibWandbLogger
