+++ src/anomalib/utils/types/__init__.py
@@ -1,0 +1,8 @@
+"""Anomalib Types."""
+
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .task_type import TaskType
+
+__all__ = ["TaskType"]
