+++ tests/legacy/pre_merge/tools/test_gradio_entrypoint.py
@@ -11,7 +11,7 @@ from typing import Callable
 
 import pytest
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.deploy import OpenVINOInferencer, TorchInferencer, export_to_openvino, export_to_torch
 from anomalib.models import Padim
 from anomalib.utils.metrics.threshold import F1AdaptiveThreshold
