+++ tests/legacy/pre_merge/tools/test_torch_entrypoint.py
@@ -11,7 +11,7 @@ from pathlib import Path
 
 import pytest
 
-from anomalib.data import TaskType
+from anomalib import TaskType
 from anomalib.deploy import export_to_torch
 from anomalib.models import Padim
 from anomalib.utils.metrics.threshold import F1AdaptiveThreshold
