+++ tests/legacy/pre_merge/utils/sweep/conftest.py
@@ -7,7 +7,8 @@ from tempfile import TemporaryDirectory
 
 import pytest
 
-from anomalib.data import MVTec, TaskType
+from anomalib import TaskType
+from anomalib.data import MVTec
 from anomalib.deploy import ExportMode
 from anomalib.engine import Engine
 from anomalib.models import Padim
